Move decorator of permissions to auth's directory
