+++ server/api/routes/comercio_route.py
@@ -5,7 +5,7 @@ from flask_jwt_extended import jwt_required, get_jwt_identity
 from server.config import db
 from server.models import Comercio, TipoResponsable, Provincia, PuntoVenta
 from server.auth.models import Usuario
-from server.api.decorators import permission_required
+from server.auth.decorators import permission_required
 
 comercio_bp = Blueprint("comercio_bp", __name__)
